입력 양식 리스너 추가

+++ app/src/main/java/com/aiden/accountwallet/data/vo/AccountInfo.kt
@@ -1,7 +1,7 @@
 package com.aiden.accountwallet.data.vo
 
 data class AccountInfo(
-    val keyIndex:Int = -1,
+    val keyIndex:Long = -1,
     val providerName:String,
     val typeIdx:Int = 0,
     val typeTag:String,
